requirements.txt
flask>=2.0
pdfplumber>=0.7
python-docx>=0.8
werkzeug>=2.0
